Version 1.1
A new interface has been created for the add-on, with the first concept of configuration options.
An error has been fixed in which the description of certain objects and the contents of some dialog boxes were not read.
Fixed a bug where the value of progress bars was not read automatically.
Fixed a bug where links contained in e-mail messages and web pages could not be focused correctly.
A problem with reading Excel cells has been fixed.
A logic has been created to check whether the read-only status is relevant, in order to avoid unnecessary announcements.