Version 1.2
Private test version, initially released as 1.1 and later updated to 1.2.
A new configuration option has been created that allows you to choose whether or not the add-on should translate the names of the types and states of the controls.
A logic of message extension levels has been implemented - long, medium and short. At the long level (default), all possible information will be spoken. At the medium level, some information will be suppressed and at the short level, only the essential information will be spoken.