Version 1.5
The “internal_link” status (which identifies links to the same page) has been added to the list of states to be announced.
A few more controls have also been added to the list of control types to be announced.
A logic has been created that checks for the presence of the old version of the add-on and removes it before installing this new version.
The add-on is now available in the NVDA add-ons store. Just search for “Pausing Information”.
Fixed an issue where Settings for the custom message extension level were lost when NVDA was restarted or when its language was changed.
The code of the add-on has been simplified and unnecessary and repeated parts have been removed to make it easier to maintain.
The add-on functionality has been integrated into DeltaTalk as “Virtual Vision Mode”, so this will be the last stand-alone version.